[IC-110] chore: TeamPolicy와 TeamPolicySet 분리(#92)
